BaseTypes slice
