+++ src/Fluent.Calculations.Primitives.Tests/Composition/Money/Money.cs
@@ -1,4 +1,4 @@
-﻿using Fluent.Calculations.Primitives;
+﻿using Fluent.Calculations.Primitives.BaseTypes;
 
 namespace Fluent.Calculations.Finance;
