+++ src/Fluent.Calculations.Primitives.Tests/Composition/Money/ToMoney.cs
@@ -1,4 +1,4 @@
-﻿using Fluent.Calculations.Primitives;
+﻿using Fluent.Calculations.Primitives.BaseTypes;
 using System.Runtime.CompilerServices;
 namespace Fluent.Calculations.Finance;
