+++ src/Fluent.Calculations.Primitives.Tests/Conditions/ConditionTests.cs
@@ -1,4 +1,5 @@
-﻿using FluentAssertions;
+﻿using Fluent.Calculations.Primitives.BaseTypes;
+using FluentAssertions;
 
 namespace Fluent.Calculations.Primitives.Tests.Conditions;
 public class ConditionTests
