+++ src/Fluent.Calculations.Primitives/BaseTypes/IName.cs
@@ -1,0 +1,6 @@
+﻿namespace Fluent.Calculations.Primitives.BaseTypes;
+
+internal interface IName
+{
+    void Set(string name);
+}
