+++ src/Fluent.Calculations.Primitives/Expressions/LamdaExpressionPrefixRemover.cs
@@ -1,4 +1,4 @@
-﻿namespace Fluent.Calculations.Primitives;
+﻿namespace Fluent.Calculations.Primitives.Expressions;
 
 internal static class LamdaExpressionPrefixRemover
 {
