--- src/Fluent.Calculations.Primitives/IName.cs
@@ -1,7 +1,0 @@
-﻿namespace Fluent.Calculations.Primitives
-{
-    internal interface IName
-    {
-        void Set(string name);
-    }
-}
