--- src/Fluent.Calculations.Primitives/Tag.cs
@@ -1,11 +1,0 @@
-﻿namespace Fluent.Calculations.Primitives
-{
-    public class Tag
-    {
-        public required string Name { get; init; }
-
-        private Tag() { }
-
-        public static Tag Create(string name) => new() { Name = name };
-    }
-}
