--- src/Fluent.Calculations.Primitives/TagsList.cs
@@ -1,15 +1,0 @@
-﻿namespace Fluent.Calculations.Primitives
-{
-    public class TagsList : List<Tag>
-    {
-        internal TagsList()
-        {
-
-        }
-        internal TagsList(IEnumerable<Tag> tags) : base(tags)
-        {
-        }
-
-        internal static TagsList Empty => new TagsList();
-    }
-}
